[QI2-1292] Seperate QI2 backend provider
